+++ src/data-sync/pull-strategy.js
@@ -1,6 +1,6 @@
 import cloneDeep from 'lodash.clonedeep';
 import isPublished from '../util/isPublished';
-import Logger from '../logger/logger';
+import logger from '../logger/logger';
 import ApiHelper from './api-helper';
 import apiHelper from "./api-helper";
 
@@ -68,7 +68,7 @@ class PullStrategy {
         });
       }
     } catch (err) {
-      Logger.log('error', err);
+      logger.error(err);
     }
 
     let screenCampaigns = [];
@@ -82,7 +82,7 @@ class PullStrategy {
         ({ campaign }) => campaign
       );
     } catch (err) {
-      Logger.log('error', err);
+      logger.error(err);
     }
 
     return new Promise((resolve) => {
@@ -189,8 +189,7 @@ class PullStrategy {
     try {
       screen = await this.apiHelper.getPath(screenPath);
     } catch (err) {
-      Logger.log(
-        'warn',
+      logger.warn(
         `Screen (${screenPath}) not loaded. Aborting content update.`
       );
 
@@ -198,41 +197,41 @@ class PullStrategy {
     }
 
     if (screen === null) {
-      Logger.log('warn', `Screen (${screenPath}) not loaded`);
+      logger.warn(`Screen (${screenPath}) not loaded`);
       return;
     }
 
     const newScreen = cloneDeep(screen);
 
-    // Campaigns data
-    let hasActiveCampaign = false;
+    newScreen.hasActiveCampaign = false;
 
     const newScreenChecksums = newScreen?.relationsChecksum ?? [];
-    const oldScreenChecksums = this.lastestScreenData?.relationsChecksum ?? null;
+    const oldScreenChecksums =
+      this.lastestScreenData?.relationsChecksum ?? null;
 
     if (
       oldScreenChecksums === null ||
       oldScreenChecksums?.campaigns !== newScreenChecksums?.campaigns ||
       oldScreenChecksums?.inScreenGroups !== newScreenChecksums?.inScreenGroups
     ) {
-      Logger.log('info', `Campaigns or screen groups modified.`);
+      logger.info(`Campaigns or screen groups modified.`);
       newScreen.campaignsData = await this.getCampaignsData(newScreen);
     } else {
-      Logger.log('info', `Campaigns or screen groups not modified.`);
+      logger.info(`Campaigns or screen groups not modified.`);
       newScreen.campaignsData = this.lastestScreenData.campaignsData;
     }
 
     if (newScreen.campaignsData.length > 0) {
       newScreen.campaignsData.forEach(({ published }) => {
         if (isPublished(published)) {
-          hasActiveCampaign = true;
+          newScreen.hasActiveCampaign = true;
         }
       });
     }
 
     // With active campaigns, we override region/layout values.
-    if (hasActiveCampaign) {
-      Logger.log('info', `Has active campaign.`);
+    if (newScreen.hasActiveCampaign) {
+      logger.info(`Has active campaign.`);
 
       // Create ulid to connect the campaign with the regions/playlists.
       const campaignRegionId = '01G112XBWFPY029RYFB8X2H4KD';
@@ -260,29 +259,33 @@ class PullStrategy {
         newScreen.regionData
       );
     } else {
+      logger.info(`Has no active campaign.`);
+
       // Get layout: Defines layout and regions.
       if (
+        this.lastestScreenData?.hasActiveCampaign ||
         oldScreenChecksums === null ||
         oldScreenChecksums?.layout !== newScreenChecksums?.layout
       ) {
-        Logger.log('info', `Layout changed since last fetch.`);
+        logger.info(`Layout changed since last fetch.`);
         newScreen.layoutData = await this.apiHelper.getPath(newScreen.layout);
       } else {
         // Get layout: Defines layout and regions.
-        Logger.log('info', `Layout not changed since last fetch.`);
+        logger.info(`Layout not changed since last fetch.`);
         newScreen.layoutData = this.lastestScreenData.layoutData;
       }
 
       // Fetch regions playlists: Yields playlists of slides for the regions
       if (
+        this.lastestScreenData?.hasActiveCampaign ||
         oldScreenChecksums === null ||
         oldScreenChecksums?.regions !== newScreenChecksums?.regions
       ) {
-        Logger.log('info', `Regions changed since last fetch.`);
+        logger.info(`Regions changed since last fetch.`);
         const regions = await this.getRegions(newScreen.regions);
         newScreen.regionData = await this.getSlidesForRegions(regions);
       } else {
-        Logger.log('info', `Regions not changed since last fetch.`);
+        logger.info(`Regions not changed since last fetch.`);
         newScreen.regionData = this.lastestScreenData.regionData;
       }
     }
@@ -353,8 +356,7 @@ class PullStrategy {
 
           // A slide cannot work without templateData. Mark as invalid.
           if (slide.templateData === null) {
-            Logger.log(
-              'warn',
+            logger.warn(
               `Template (${slide.templateInfo['@id']}) not loaded, slideId: ${slide['@id']}`
             );
             slide.invalid = true;
@@ -395,8 +397,6 @@ class PullStrategy {
       }
     }
     /* eslint-enable no-restricted-syntax,no-await-in-loop */
-
-    Logger.log('info', `Emitting screen data.`);
 
     this.lastestScreenData = newScreen;