+++ src/remote-component.config.js
@@ -1,9 +1,7 @@
-/* eslint-disable global-require */
-/**
- * Dependencies for Remote Components
- */
-module.exports = {
-  resolve: {
-    react: require('react'),
-  },
+import react from 'react';
+
+const resolve = {
+  react,
 };
+
+export default resolve;