+++ src/service/content-service.js
@@ -1,16 +1,16 @@
 import sha256 from 'crypto-js/sha256';
 import Base64 from 'crypto-js/enc-base64';
-import cloneDeep from 'lodash.clonedeep';
 import { object, string } from 'prop-types';
 import Logger from '../logger/logger';
-import DataSync from '../data-sync/data-sync';
-import ScheduleService from './scheduleService';
-import ConfigLoader from '../config-loader';
 import PullStrategy from '../data-sync/pull-strategy';
 import {
   screenForPlaylistPreview,
   screenForSlidePreview,
 } from '../util/preview';
+import logger from '../logger/logger';
+import DataSync from '../data-sync/data-sync';
+import ScheduleService from './schedule-service';
+import ConfigLoader from '../util/config-loader';
 
 /**
  * ContentService.
@@ -48,7 +48,7 @@ class ContentService {
    * @param {string} screenPath Path to the screen.
    */
   startSyncing(screenPath) {
-    Logger.log('info', 'Starting data synchronization');
+    logger.info('Starting data synchronization');
 
     ConfigLoader.loadConfig().then((config) => {
       const dataStrategyConfig = { ...config.dataStrategy.config };
@@ -68,10 +68,10 @@ class ContentService {
    * Stop sync event handler.
    */
   stopSyncHandler() {
-    Logger.log('info', 'Event received: Stop data synchronization');
+    logger.info('Event received: Stop data synchronization');
 
     if (this.dataSync) {
-      Logger.log('info', 'Stopping data synchronization');
+      logger.info('Stopping data synchronization');
       this.dataSync.stop();
       this.dataSync = null;
     }
@@ -93,6 +93,9 @@ class ContentService {
       `Event received: Start data synchronization from ${data?.screenPath}`
     );
     if (data?.screenPath) {
+      logger.info(
+        `Event received: Start data synchronization from ${data.screenPath}`
+      );
       this.startSyncing(data.screenPath);
     } else {
       Logger.log('error', 'Error: screenPath not set.');
@@ -106,7 +109,7 @@ class ContentService {
    *   The event.
    */
   contentHandler(event) {
-    Logger.log('info', 'Event received: content');
+    logger.info('Event received: content');
 
     const data = event.detail;
     this.currentScreen = data.screen;
@@ -123,11 +126,11 @@ class ContentService {
     // TODO: Handle issue where region data is not present for a given region. Remove given region content.
 
     if (newHash !== this.screenHash) {
-      Logger.log('info', 'Screen has changed. Emitting screen.');
+      logger.info('Screen has changed. Emitting screen.');
       this.screenHash = newHash;
       ContentService.emitScreen(screenData);
     } else {
-      Logger.log('info', 'Screen has not changed. Not emitting screen.');
+      logger.info('Screen has not changed. Not emitting screen.');
 
       // eslint-disable-next-line guard-for-in,no-restricted-syntax
       for (const regionKey in data.screen.regionData) {
@@ -147,7 +150,7 @@ class ContentService {
     const data = event.detail;
     const regionId = data.id;
 
-    Logger.log('info', `Event received: regionReady for ${regionId}`);
+    logger.info(`Event received: regionReady for ${regionId}`);
 
     if (this.currentScreen) {
       this.scheduleService.updateRegion(
@@ -167,7 +170,7 @@ class ContentService {
     const data = event.detail;
     const regionId = data.id;
 
-    Logger.log('info', `Event received: regionRemoved for ${regionId}`);
+    logger.info(`Event received: regionRemoved for ${regionId}`);
 
     this.scheduleService.regionRemoved(regionId);
   }
@@ -176,7 +179,7 @@ class ContentService {
    * Start the engine.
    */
   start() {
-    Logger.log('info', 'Content service started.');
+    logger.info('Content service started.');
 
     document.addEventListener('stopDataSync', this.stopSyncHandler);
     document.addEventListener('startDataSync', this.startDataSyncHandler);
@@ -190,7 +193,7 @@ class ContentService {
    * Stop the engine.
    */
   stop() {
-    Logger.log('info', 'Content service stopped.');
+    logger.info('Content service stopped.');
 
     document.removeEventListener('stopDataSync', this.stopSyncHandler);
     document.removeEventListener('startDataSync', this.startDataSyncHandler);
@@ -293,7 +296,7 @@ class ContentService {
    *   Screen data.
    */
   static emitScreen(screen) {
-    Logger.log('info', 'Emitting screen');
+    logger.info('Emitting screen');
 
     const event = new CustomEvent('screen', {
       detail: {