+++ src/service/schedule-service.js
@@ -1,11 +1,11 @@
 import cloneDeep from 'lodash.clonedeep';
 import sha256 from 'crypto-js/sha256';
 import Md5 from 'crypto-js/md5';
-import RRule from 'rrule';
 import Base64 from 'crypto-js/enc-base64';
 import isPublished from '../util/isPublished';
-import Logger from '../logger/logger';
-import ConfigLoader from '../config-loader';
+import logger from '../logger/logger';
+import ConfigLoader from '../util/config-loader';
+import ScheduleUtils from '../util/schedule';
 
 /**
  * ScheduleService.
@@ -27,7 +27,7 @@ class ScheduleService {
   }
 
   checkForEmptyContent() {
-    Logger.log('info', 'Checking for empty content.');
+    logger.info('Checking for empty content.');
 
     // Check for empty content.
     const values = Object.values(this.regions);
@@ -52,12 +52,15 @@ class ScheduleService {
    * @param {string} regionId - The region id.
    */
   regionRemoved(regionId) {
-    Logger.log('info', `removing scheduling interval for region: ${regionId}`);
+    logger.info(`removing scheduling interval for region: ${regionId}`);
 
     if (Object.prototype.hasOwnProperty.call(this.intervals, regionId)) {
       clearInterval(this.intervals[regionId]);
       delete this.intervals[regionId];
     }
+
+    // Remove cached version of region data.
+    delete this.regions[regionId];
   }
 
   /**
@@ -67,10 +70,10 @@ class ScheduleService {
    * @param {object} region - The region content, with playlists and slides, to start scheduling.
    */
   updateRegion(regionId, region) {
-    Logger.log('info', `ScheduleService: updateRegion(${regionId})`);
+    logger.info(`ScheduleService: updateRegion(${regionId})`);
 
     if (!region || !regionId) {
-      Logger.log('info', `ScheduleService: regionId and/or region not set.`);
+      logger.info(`ScheduleService: regionId and/or region not set.`);
       return;
     }
 
@@ -96,8 +99,7 @@ class ScheduleService {
 
         // Extra check because of async.
         if (!Object.prototype.hasOwnProperty.call(intervals, regionId)) {
-          Logger.log(
-            'info',
+          logger.info(
             `registering scheduling interval for region: ${regionId}, with an update rate of ${schedulingInterval}`
           );
 
@@ -121,7 +123,7 @@ class ScheduleService {
    * @param {string} regionId - The region to check.
    */
   checkScheduling(regionId) {
-    Logger.log('info', `checkScheduling for region: ${regionId}`);
+    logger.info(`checkScheduling for region: ${regionId}`);
 
     const region = this.regions[regionId];
 
@@ -153,7 +155,7 @@ class ScheduleService {
    *   Array of slides.
    */
   sendSlides(regionId, slides) {
-    Logger.log('info', `sendSlides regionContent-${regionId}`);
+    logger.info(`sendSlides regionContent-${regionId}`);
     const event = new CustomEvent(`regionContent-${regionId}`, {
       detail: {
         slides,
@@ -174,10 +176,6 @@ class ScheduleService {
   static findScheduledSlides(playlists, regionId) {
     const slides = [];
 
-    const now = new Date();
-    const startOfDay = new Date();
-    startOfDay.setUTCHours(0, 0, 0, 0);
-
     playlists.forEach((playlist) => {
       const { schedules } = playlist;
 
@@ -185,38 +183,32 @@ class ScheduleService {
         return;
       }
 
-      let occurs = true;
+      let active = true;
 
       // If schedules are set for the playlist, do not show playlist unless a schedule is active.
       if (schedules.length > 0) {
-        occurs = false;
+        active = false;
 
-        schedules.forEach((schedule) => {
-          const rrule = RRule.fromString(schedule.rrule.replace('\\n', '\n'));
-          rrule.between(
-            // Subtract duration from now to make sure all relevant occurrences are considered.
-            new Date(
-              now.getTime() - (schedule.duration ? schedule.duration * 1000 : 0)
-            ),
-            now,
-            true,
-            function iterator(occurrenceDate) {
-              const occurrenceEnd = new Date(
-                occurrenceDate.getTime() + schedule.duration * 1000
-              );
-
-              if (now >= occurrenceDate && now <= occurrenceEnd) {
-                occurs = true;
-                // Break the iteration.
-                return false;
-              }
-              return true;
-            }
+        // Run through all schedule item and see if it occurs now. If one or more occur now, the playlist is active.
+        schedules.every((schedule) => {
+          const scheduleOccurs = ScheduleUtils.occursNow(
+            schedule.rrule,
+            schedule.duration
           );
+
+          if (scheduleOccurs) {
+            active = true;
+
+            // Break iteration.
+            return false;
+          }
+
+          // Continue iteration.
+          return true;
         });
       }
 
-      if (occurs) {
+      if (active) {
         playlist?.slidesData?.forEach((slide) => {
           if (!isPublished(slide.published)) {
             return;
@@ -229,6 +221,8 @@ class ScheduleService {
           newSlide.executionId = `EXE-ID-${executionId}`;
           slides.push(newSlide);
         });
+      } else {
+        logger.log('info', `Playlist ${playlist['@id']} not scheduled for now`);
       }
     });