+++ CHANGELOG.md
@@ -5,6 +5,51 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - Added support for previews.
+- [#140](https://github.com/os2display/display-client/pull/140)
+  - Fixed issue where campaign end resulted in blank screen.
+
+## [2.1.1] - 2024-10-23
+
+- [#138](https://github.com/os2display/display-client/pull/138)
+  - Avoided setting ER105 in url when no token exists in local storage.
+- [#137](https://github.com/os2display/display-client/pull/137)
+  - Add `no-cache´ directive to nginx setup.
+
+## [2.1.0] - 2024-10-23
+
+- [#135](https://github.com/os2display/display-client/pull/135)
+  - Fixed cursor being hidden when not in debug mode.
+  - Moved registration of listeners in useEffect.
+- [#134](https://github.com/os2display/display-client/pull/134)
+  - Fixed remote loader for touch regions.
+- [#133](https://github.com/os2display/display-client/pull/133)
+  - Added error message ER201 on screen when remote component could not load.
+  - Added error timestamp to remote component loader, to force reload on error.
+- [#132](https://github.com/os2display/display-client/pull/132)
+  - Remove token errors after re-login.
+- [#131](https://github.com/os2display/display-client/pull/131)
+  - Moved localstorage calls into helper class AppStorage.
+  - Cleaned up project structure.
+  - Fixed async flow for release check.
+  - Added previous boot timestamp (pb) to url.
+  - Try to refresh token on reauthenticate event.
+  - Added status and error to url.
+  - Added error code to login page, if set.
+- [#130](https://github.com/os2display/display-client/pull/130)
+  - Fixed issue with svg loading.
+- [#127](https://github.com/os2display/display-client/pull/127)
+  - Migrate from CRA to Vite
+  - Upgrade node from 14 -> 20
+  - Winston logging to Pino logging
+  - Set infrastructure node version to 20.
+  - Fixed base path to /client.
+- [#118](https://github.com/os2display/display-client/pull/118)
+  - Added api url to localstorage.
+
+## [2.0.4] - 2024-08-14
+
+- [#128](https://github.com/os2display/display-client/pull/128)
+  - Fixed rrule evaluation to handle local time correctly.
 
 ## [2.0.3] - 2024-05-21
 
@@ -34,7 +79,8 @@ All notable changes to this project will be documented in this file.
 
 - [#119](https://github.com/os2display/display-client/pull/119)
   - Changed to v2 routing.
-- Limited the number of API calls by comparing relationsChecksum.
+- [#117](https://github.com/os2display/display-client/pull/117)
+  - Limited the number of API calls by comparing relationsChecksum.
 
 ## [1.3.5] - 2023-09-14